refactor: 変数名の変更
